move sub-modules to top level access. create initial cost model draft
