+++ electrolyzer/__init__.py
@@ -3,3 +3,7 @@
 __author__ = """Christopher Bay"""
 __email__ = "christopher.bay@nrel.gov"
 __version__ = "0.1.0"
+
+from .electrolyzer import *
+from .electrolyzer_supervisor import *
+from .electrolyzer_cost import *
